split CardEditor - final step
