+++ colab-webapp/src/main/node/app/src/components/cards/CardEditor.tsx
@@ -5,118 +5,61 @@
  * Licensed under the MIT License
  */
 
-import { css, cx } from '@emotion/css';
+import { css } from '@emotion/css';
 import { Card, CardContent } from 'colab-rest-client';
 import * as React from 'react';
 import { CirclePicker } from 'react-color';
 import { ReflexContainer, ReflexElement, ReflexSplitter } from 'react-reflex';
-import 'react-reflex/styles.css';
-import { useNavigate } from 'react-router-dom';
 import * as API from '../../API/api';
 import useTranslations from '../../i18n/I18nContext';
-import { useAppDispatch, useLoadingState } from '../../store/hooks';
+import { useAppDispatch } from '../../store/hooks';
 import { useCardACLForCurrentUser } from '../../store/selectors/aclSelector';
-import { useAndLoadSubCards, useVariantsOrLoad } from '../../store/selectors/cardSelector';
+import { useAndLoadSubCards } from '../../store/selectors/cardSelector';
 import { useAndLoadIfOnlyEmptyDocuments } from '../../store/selectors/documentSelector';
 import { space_md, space_sm } from '../../styling/style';
 import { cardColors } from '../../styling/theme';
-import IconButton from '../common/element/IconButton';
-import ConfirmDeleteOpenCloseModal from '../common/layout/ConfirmDeleteModal';
 import Flex from '../common/layout/Flex';
 import Icon from '../common/layout/Icon';
-import {
-  Item,
-  SideCollapsibleCtx,
-  SideCollapsibleMenu,
-  SideCollapsiblePanelBody,
-} from '../common/layout/SideCollapsiblePanel';
+import { Item, SideCollapsibleCtx } from '../common/layout/SideCollapsibleContext';
 import { DocumentOwnership } from '../documents/documentCommonType';
-import DocEditorToolbox, {
-  defaultDocEditorContext,
-  DocEditorCtx,
-} from '../documents/DocumentEditorToolbox';
-import DocumentList from '../documents/DocumentList';
-import { ResourceAndRef, ResourceOwnership } from '../resources/resourcesCommonType';
-import {
-  ResourcesCtx,
-  ResourcesMainViewHeader,
-  ResourcesMainViewPanel,
-} from '../resources/ResourcesMainView';
+import { ResourcesMainViewHeader, ResourcesMainViewPanel } from '../resources/ResourcesMainView';
 import CardAssignmentsPanel from '../team/CardAssignments';
-import CardThumb from './CardThumb';
-import Dndwrapper from './dnd/Dndwrapper';
-import { ProgressBar, ProgressBarEditor } from './ProgressBar';
-import { computeNav } from './VariantSelector';
+import CardEditorDeliverable from './CardEditorDeliverable';
+import CardEditorHeader from './CardEditorHeader';
+import CardEditorSideMenu from './CardEditorSideMenu';
+import CardEditorSidePanel from './CardEditorSidePanel';
+import CardEditorSubCards from './CardEditorSubCards';
 
 interface CardEditorProps {
   card: Card;
-  variant: CardContent;
+  cardContent: CardContent;
+  readOnly?: boolean;
 }
 
-export default function CardEditor({ card, variant }: CardEditorProps): JSX.Element {
+export default function CardEditor({ card, cardContent }: CardEditorProps): JSX.Element {
   const i18n = useTranslations();
   const dispatch = useAppDispatch();
-  const navigate = useNavigate();
-  //const location = useLocation();
 
-  //const { currentUser } = useCurrentUser();
-
-  const variants = useVariantsOrLoad(card) || [];
-  const hasVariants = variants.length > 1 && variant != null;
-  //const variantNumber = hasVariants ? variants.indexOf(variant) + 1 : undefined;
-
-  const contents = useVariantsOrLoad(card);
-
-  const subCards = useAndLoadSubCards(variant.id);
-
-  const variantPager = computeNav(contents, variant.id);
+  const subCards = useAndLoadSubCards(cardContent.id);
 
   const { canRead, canWrite } = useCardACLForCurrentUser(card.id);
-  const readOnly = !canWrite || variant.frozen;
-  //const [showTypeDetails, setShowTypeDetails] = React.useState(false);
-  const [selectedDocId, setSelectedDocId] = React.useState<number | null>(null);
-  const [lastCreatedDocId, setLastCreatedDocId] = React.useState<number | null>(null);
-  const [editMode, setEditMode] = React.useState(defaultDocEditorContext.editMode);
-  const [showTree, setShowTree] = React.useState(false);
-  const [markDownMode, setMarkDownMode] = React.useState(false);
-  const [editToolbar, setEditToolbar] = React.useState(defaultDocEditorContext.editToolbar);
+  const readOnly = !canWrite || cardContent.frozen;
+
   const [openKey, setOpenKey] = React.useState<string | undefined>(undefined);
-
-  const [selectedResource, selectResource] = React.useState<ResourceAndRef | null>(null);
-  const [lastCreatedResourceId, setLastCreatedResourceId] = React.useState<number | null>(null);
-
-  const TXToptions = {
-    showTree: showTree,
-    setShowTree: setShowTree,
-    markDownMode: markDownMode,
-    setMarkDownMode: setMarkDownMode,
-  };
-  const { isLoading, startLoading, stopLoading } = useLoadingState();
 
   const hasNoSubCard = !subCards || subCards.length < 1;
 
   const deliverableDocContext: DocumentOwnership = {
     kind: 'DeliverableOfCardContent',
-    ownerId: variant.id!,
+    ownerId: cardContent.id!,
   };
 
   const { empty: hasNoDeliverableDoc } = useAndLoadIfOnlyEmptyDocuments(deliverableDocContext);
 
-  const resourceOwnership: ResourceOwnership = {
-    kind: 'CardOrCardContent',
-    cardId: card.id || undefined,
-    cardContentId: variant.id,
-    hasSeveralVariants: hasVariants,
-  };
-
-  const sideBarItems: Record<string, Item> = {
-    resources: {
+  const sideBarItems: Item[] = [
+    {
+      key: 'resources',
       icon: 'menu_book',
-      // nextToIconElement: (
-      //   <div className={text_sm}>
-      //     <ResourcesListNb resourcesOwnership={resourceOwnership} />
-      //   </div>
-      // ),
       title: i18n.modules.resource.documentation,
       header: (
         <ResourcesMainViewHeader
@@ -132,7 +75,8 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
       ),
       className: css({ overflow: 'auto' }),
     },
-    team: {
+    {
+      key: 'team',
       icon: 'group',
       title: i18n.team.assignment.labels.assignments,
       children: (
@@ -142,7 +86,8 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
       ),
       className: css({ overflow: 'auto' }),
     },
-    colors: {
+    {
+      key: 'colors',
       icon: 'palette',
       title: i18n.modules.card.settings.color,
       children: (
@@ -167,35 +112,15 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
         />
       ),
     },
-  };
-
-  //const { stickyNotesForDest } = useStickyNoteLinksForDest(card.id);
-  /*  const closeRouteCb = React.useCallback(
-    (route: string) => {
-      navigate(location.pathname.replace(new RegExp(route + '$'), ''));
-    },
-    [location.pathname, navigate],
-  ); */
-
-  /* const goto = React.useCallback(
-    (card: Card, variant: CardContent) => {
-      navigate(`../card/${card.id}/v/${variant.id}`);
-    },
-    [navigate],
-  ); */
+  ];
 
   if (card.id == null) {
     return <i>{i18n.modules.card.error.withoutId}</i>;
   } else {
     return (
       <Flex direction="column" grow={1} align="stretch">
-        <Flex direction="column" align="stretch">
-          {readOnly ? (
-            <ProgressBar card={card} variant={variant} tall />
-          ) : (
-            <ProgressBarEditor card={card} variant={variant} />
-          )}
-        </Flex>
+        <CardEditorHeader card={card} cardContent={cardContent} readOnly={readOnly} />
+
         <Flex direction="row" grow={1} align="stretch" className={css({ overflow: 'auto' })}>
           <SideCollapsibleCtx.Provider
             value={{
@@ -217,68 +142,13 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
                     minSize={65}
                     flex={hasNoSubCard ? 1 : hasNoDeliverableDoc ? 0 : 0.5}
                   >
-                    <Flex
-                      grow={1}
-                      direction="column"
-                      align="stretch"
-                      className={cx(css({ overflow: 'auto' }))}
-                    >
-                      <Flex grow={1} align="stretch" className={css({ overflow: 'hidden' })}>
-                        <DocEditorCtx.Provider
-                          value={{
-                            selectedDocId,
-                            setSelectedDocId,
-                            lastCreatedId: lastCreatedDocId,
-                            setLastCreatedId: setLastCreatedDocId,
-                            editMode,
-                            setEditMode,
-                            editToolbar,
-                            setEditToolbar,
-                            TXToptions,
-                          }}
-                        >
-                          <Flex direction="column" grow={1} align="stretch">
-                            <Flex
-                              direction="column"
-                              grow={1}
-                              className={css({
-                                overflow: 'auto',
-                              })}
-                              align="stretch"
-                            >
-                              <Flex direction="column" align="stretch">
-                                {!readOnly && variant.id && (
-                                  <DocEditorToolbox
-                                    open={true}
-                                    docOwnership={deliverableDocContext}
-                                  />
-                                )}
-                              </Flex>
-                              <Flex
-                                direction="column"
-                                grow={1}
-                                align="stretch"
-                                className={css({ overflow: 'auto', padding: space_sm })}
-                              >
-                                {canRead != undefined &&
-                                  (canRead ? (
-                                    variant.id ? (
-                                      <DocumentList
-                                        docOwnership={deliverableDocContext}
-                                        readOnly={readOnly}
-                                      />
-                                    ) : (
-                                      <span>{i18n.modules.card.infos.noDeliverable}</span>
-                                    )
-                                  ) : (
-                                    <span>{i18n.httpErrorMessage.ACCESS_DENIED}</span>
-                                  ))}
-                              </Flex>
-                            </Flex>
-                          </Flex>
-                        </DocEditorCtx.Provider>
-                      </Flex>
-                    </Flex>
+                    {/* ******************************** DELIVERABLE ******************************** */}
+                    <CardEditorDeliverable
+                      card={card}
+                      cardContent={cardContent}
+                      readOnly={readOnly}
+                    />
+                    {/* ***************************************************************************** */}
                   </ReflexElement>
                   <ReflexSplitter
                     className={css({
@@ -301,21 +171,12 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
                     resizeWidth={false}
                     minSize={42}
                   >
-                    <Flex className={css({ width: '100%', overflow: 'auto', flexGrow: 1 })}>
-                      <Dndwrapper cards={subCards}>
-                        <CardThumb
-                          card={card}
-                          variant={variant}
-                          variants={variants}
-                          showSubcards={true}
-                          depth={2}
-                          mayOrganize={true}
-                          showPreview={false}
-                          withoutHeader={true}
-                          coveringColor={false}
-                        />
-                      </Dndwrapper>
-                    </Flex>
+                    {/* ******************************** SUB CARDS ******************************** */}
+                    <CardEditorSubCards
+                      card={card}
+                      cardContent={cardContent} /* readOnly={readOnly} */
+                    />
+                    {/* *************************************************************************** */}
                   </ReflexElement>
                 </ReflexContainer>
               </ReflexElement>
@@ -339,57 +200,18 @@ export default function CardEditor({ card, variant }: CardEditorProps): JSX.Elem
                 minSize={20}
                 flex={0.2}
               >
-                <ResourcesCtx.Provider
-                  value={{
-                    resourceOwnership,
-                    selectedResource,
-                    selectResource,
-                    lastCreatedId: lastCreatedResourceId,
-                    setLastCreatedId: setLastCreatedResourceId,
-                  }}
-                >
-                  <SideCollapsiblePanelBody className={css({ overflow: 'hidden' })} />
-                </ResourcesCtx.Provider>
+                {/* ******************************** SIDE PANEL ******************************** */}
+                <CardEditorSidePanel
+                  card={card}
+                  cardContent={cardContent} /* readOnly={readOnly} */
+                />
+                {/* **************************************************************************** */}
               </ReflexElement>
             </ReflexContainer>
-            <Flex
-              direction="column"
-              justify="space-between"
-              className={css({
-                borderLeft: '1px solid var(--divider-main)',
-              })}
-            >
-              <SideCollapsibleMenu defaultOpenKey="resources" />
 
-              <ConfirmDeleteOpenCloseModal
-                buttonLabel={
-                  <IconButton
-                    icon={'delete'}
-                    title={i18n.modules.content.deleteBlock}
-                    onClick={() => {}}
-                    className={cx(css({ color: 'var(--error-main)' }))}
-                  />
-                }
-                confirmButtonLabel={i18n.modules.card.deleteCardVariant(hasVariants)}
-                message={<p>{i18n.modules.card.confirmDeleteCardVariant(hasVariants)}</p>}
-                onConfirm={() => {
-                  startLoading();
-                  if (hasVariants) {
-                    dispatch(API.deleteCardContent(variant)).then(() => {
-                      navigate(`../card/${card.id}/v/${variantPager?.next.id}`);
-                      stopLoading();
-                    });
-                  } else {
-                    dispatch(API.deleteCard(card)).then(() => {
-                      navigate('../');
-                      stopLoading();
-                    });
-                  }
-                }}
-                title={i18n.modules.card.deleteCardVariant(hasVariants)}
-                isConfirmButtonLoading={isLoading}
-              />
-            </Flex>
+            {/* ******************************** SIDE MENU ******************************** */}
+            <CardEditorSideMenu card={card} cardContent={cardContent} /* readOnly={readOnly} */ />
+            {/* *************************************************************************** */}
           </SideCollapsibleCtx.Provider>
         </Flex>
       </Flex>
