+++ colab-webapp/src/main/node/app/src/components/cards/CardEditorHeader.tsx
@@ -4,6 +4,7 @@
  *
  * Licensed under the MIT License
  */
+
 import { css } from '@emotion/css';
 import { Card, CardContent, entityIs } from 'colab-rest-client';
 import * as React from 'react';
@@ -12,9 +13,7 @@ import { Route, Routes, useLocation, useNavigate } from 'react-router-dom';
 import * as API from '../../API/api';
 import useTranslations from '../../i18n/I18nContext';
 import { useAppDispatch } from '../../store/hooks';
-import { useCardACLForCurrentUser } from '../../store/selectors/aclSelector';
-import { useAndLoadSubCards, useVariantsOrLoad } from '../../store/selectors/cardSelector';
-import { useAndLoadIfOnlyEmptyDocuments } from '../../store/selectors/documentSelector';
+import { useVariantsOrLoad } from '../../store/selectors/cardSelector';
 import { useCurrentUser } from '../../store/selectors/userSelector';
 import { heading_sm, lightIconButtonStyle, space_sm } from '../../styling/style';
 import { cardColors } from '../../styling/theme';
@@ -24,17 +23,21 @@ import DropDownMenu from '../common/layout/DropDownMenu';
 import Flex from '../common/layout/Flex';
 import Icon from '../common/layout/Icon';
 import Modal from '../common/layout/Modal';
-import { DocumentOwnership } from '../documents/documentCommonType';
 import CardSettings from './CardSettings';
+import { ProgressBar, ProgressBarEditor } from './ProgressBar';
 import StatusDropDown from './StatusDropDown';
 import { VariantPager } from './VariantSelector';
 
-interface CardEditorProps {
+interface CardEditorHeaderProps {
   card: Card;
-  variant: CardContent;
+  cardContent: CardContent;
+  readOnly?: boolean;
 }
-
-export default function CardEditorHeader({ card, variant }: CardEditorProps): JSX.Element {
+export default function CardEditorHeader({
+  card,
+  cardContent,
+  readOnly,
+}: CardEditorHeaderProps): JSX.Element {
   const i18n = useTranslations();
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
@@ -43,25 +46,8 @@ export default function CardEditorHeader({ card, variant }: CardEditorProps): JS
   const { currentUser } = useCurrentUser();
 
   const variants = useVariantsOrLoad(card) || [];
-  const hasVariants = variants.length > 1 && variant != null;
-  const variantNumber = hasVariants ? variants.indexOf(variant) + 1 : undefined;
-
-  /* const contents = useVariantsOrLoad(card);*/
-
-  const subCards = useAndLoadSubCards(variant.id);
-
-  const { canWrite } = useCardACLForCurrentUser(card.id);
-  const readOnly = !canWrite || variant.frozen;
-  //const [showTypeDetails, setShowTypeDetails] = React.useState(false);
-
-  const hasNoSubCard = !subCards || subCards.length < 1;
-
-  const deliverableDocContext: DocumentOwnership = {
-    kind: 'DeliverableOfCardContent',
-    ownerId: variant.id!,
-  };
-
-  const { empty: hasNoDeliverableDoc } = useAndLoadIfOnlyEmptyDocuments(deliverableDocContext);
+  const hasVariants = variants.length > 1 && cardContent != null;
+  const variantNumber = hasVariants ? variants.indexOf(cardContent) + 1 : undefined;
 
   const closeRouteCb = React.useCallback(
     (route: string) => {
@@ -71,161 +57,174 @@ export default function CardEditorHeader({ card, variant }: CardEditorProps): JS
   );
 
   const goto = React.useCallback(
-    (card: Card, variant: CardContent) => {
-      navigate(`../card/${card.id}/v/${variant.id}`);
+    (card: Card, cardContent: CardContent) => {
+      navigate(`../card/${card.id}/v/${cardContent.id}`);
     },
     [navigate],
   );
 
   return (
-    <Flex
-      justify="space-between"
-      className={css({
-        alignItems: 'center',
-        padding: '0 ' + space_sm,
-        backgroundColor: `${card.color || cardColors.white}`,
-      })}
-    >
-      <Flex align="center">
-        <DiscreetInput
-          value={card.title || ''}
-          placeholder={i18n.modules.card.untitled}
-          readOnly={readOnly}
-          onChange={newValue => dispatch(API.updateCard({ ...card, title: newValue }))}
-          inputDisplayClassName={heading_sm}
-          autoWidth={true}
-        />
-        {hasVariants && (
-          <>
-            <span>&#xFE58;</span>
-            <DiscreetInput
-              value={
-                variant.title && variant.title.length > 0
-                  ? variant.title
-                  : i18n.modules.card.variant + `${variantNumber}`
-              }
-              placeholder={i18n.modules.content.untitled}
-              readOnly={readOnly}
-              onChange={newValue =>
-                dispatch(API.updateCardContent({ ...variant, title: newValue }))
+    <>
+      <Flex
+        justify="space-between"
+        className={css({
+          alignItems: 'center',
+          padding: '0 ' + space_sm,
+          backgroundColor: `${card.color || cardColors.white}`,
+        })}
+      >
+        <Flex align="center">
+          <DiscreetInput
+            value={card.title || ''}
+            placeholder={i18n.modules.card.untitled}
+            readOnly={readOnly}
+            onChange={newValue => dispatch(API.updateCard({ ...card, title: newValue }))}
+            inputDisplayClassName={heading_sm}
+            autoWidth={true}
+          />
+          {hasVariants && (
+            <>
+              <span>&#xFE58;</span>
+              <DiscreetInput
+                value={
+                  cardContent.title && cardContent.title.length > 0
+                    ? cardContent.title
+                    : i18n.modules.card.variant + `${variantNumber}`
+                }
+                placeholder={i18n.modules.content.untitled}
+                readOnly={readOnly}
+                onChange={newValue =>
+                  dispatch(API.updateCardContent({ ...cardContent, title: newValue }))
+                }
+              />
+              <VariantPager allowCreation={!readOnly} card={card} current={cardContent} />
+            </>
+          )}
+          <IconButton
+            icon={cardContent.frozen ? 'lock' : 'lock_open'}
+            title={i18n.modules.card.infos.cardLocked}
+            color={'var(--gray-400)'}
+            onClick={() =>
+              dispatch(API.updateCardContent({ ...cardContent, frozen: !cardContent.frozen }))
+            }
+            kind="ghost"
+            className={css({ padding: space_sm, background: 'none' })}
+          />
+          <StatusDropDown
+            value={cardContent.status}
+            readOnly={readOnly}
+            onChange={status => dispatch(API.updateCardContent({ ...cardContent, status }))}
+            kind="outlined"
+          />
+        </Flex>
+
+        {/* View mode btn *********************************************** */}
+
+        <Flex align="center">
+          <IconButton
+            title="contentOnly"
+            icon={'subtitles'}
+            kind="ghost"
+            iconSize="xs"
+            className={css({ marginRight: space_sm })}
+          ></IconButton>
+          <IconButton
+            title="Splitted"
+            icon={'space_dashboard'}
+            kind="ghost"
+            iconSize="xs"
+            className={css({ marginRight: space_sm })}
+          ></IconButton>
+          <IconButton
+            title="cardsOnly"
+            icon={'iframe'}
+            kind="ghost"
+            iconSize="xs"
+            className={css({ marginRight: space_sm })}
+          ></IconButton>
+
+          {/* handle modal routes*/}
+          <Routes>
+            <Route
+              path="settings"
+              element={
+                <Modal
+                  title={i18n.modules.card.settings.title}
+                  onClose={() => closeRouteCb('settings')}
+                  showCloseButton
+                  modalBodyClassName={css({ overflowY: 'visible' })}
+                >
+                  {closeModal => (
+                    <CardSettings onClose={closeModal} card={card} variant={cardContent} />
+                  )}
+                </Modal>
               }
             />
-            <VariantPager allowCreation={!readOnly} card={card} current={variant} />
-          </>
-        )}
-        <IconButton
-          icon={variant.frozen ? 'lock' : 'lock_open'}
-          title={i18n.modules.card.infos.cardLocked}
-          color={'var(--gray-400)'}
-          onClick={() => dispatch(API.updateCardContent({ ...variant, frozen: !variant.frozen }))}
-          kind="ghost"
-          className={css({ padding: space_sm, background: 'none' })}
-        />
-        <StatusDropDown
-          value={variant.status}
-          readOnly={readOnly}
-          onChange={status => dispatch(API.updateCardContent({ ...variant, status }))}
-          kind="outlined"
-        />
-      </Flex>
-
-      {/* View mode btn *********************************************** */}
-
-      <Flex align="center">
-        <IconButton
-          title="contentOnly"
-          icon={'subtitles'}
-          kind="ghost"
-          iconSize="xs"
-          className={css({ marginRight: space_sm })}
-        ></IconButton>
-        <IconButton
-          title="Splitted"
-          icon={'space_dashboard'}
-          kind="ghost"
-          iconSize="xs"
-          className={css({ marginRight: space_sm })}
-        ></IconButton>
-        <IconButton
-          title="cardsOnly"
-          icon={'iframe'}
-          kind="ghost"
-          iconSize="xs"
-          className={css({ marginRight: space_sm })}
-        ></IconButton>
-
-        {/* handle modal routes*/}
-        <Routes>
-          <Route
-            path="settings"
-            element={
-              <Modal
-                title={i18n.modules.card.settings.title}
-                onClose={() => closeRouteCb('settings')}
-                showCloseButton
-                modalBodyClassName={css({ overflowY: 'visible' })}
-              >
-                {closeModal => <CardSettings onClose={closeModal} card={card} variant={variant} />}
-              </Modal>
-            }
-          />
-        </Routes>
-        <DropDownMenu
-          icon={'more_vert'}
-          valueComp={{ value: '', label: '' }}
-          buttonClassName={lightIconButtonStyle}
-          entries={[
-            ...(currentUser?.admin && card.cardTypeId == null
-              ? [
-                  {
-                    value: 'createType',
-                    label: (
-                      <>
-                        <Icon icon={'account_tree'} />
-                        {i18n.modules.card.action.createAType}
-                      </>
-                    ),
-                    action: () => {
-                      dispatch(API.createCardCardType(card.id!));
+          </Routes>
+          <DropDownMenu
+            icon={'more_vert'}
+            valueComp={{ value: '', label: '' }}
+            buttonClassName={lightIconButtonStyle}
+            entries={[
+              ...(currentUser?.admin && card.cardTypeId == null
+                ? [
+                    {
+                      value: 'createType',
+                      label: (
+                        <>
+                          <Icon icon={'account_tree'} />
+                          {i18n.modules.card.action.createAType}
+                        </>
+                      ),
+                      action: () => {
+                        dispatch(API.createCardCardType(card.id!));
+                      },
                     },
-                  },
-                ]
-              : []),
-            ...(currentUser?.admin && card.cardTypeId != null
-              ? [
-                  {
-                    value: 'removeType',
-                    label: (
-                      <>
-                        <Icon icon={'eco'} /> {i18n.modules.card.action.removeTheType}
-                      </>
-                    ),
-                    action: () => {
-                      dispatch(API.removeCardCardType(card.id!));
+                  ]
+                : []),
+              ...(currentUser?.admin && card.cardTypeId != null
+                ? [
+                    {
+                      value: 'removeType',
+                      label: (
+                        <>
+                          <Icon icon={'eco'} /> {i18n.modules.card.action.removeTheType}
+                        </>
+                      ),
+                      action: () => {
+                        dispatch(API.removeCardCardType(card.id!));
+                      },
                     },
-                  },
-                ]
-              : []),
-            {
-              value: 'createVariant',
-              label: (
-                <>
-                  <Icon icon={'library_add'} /> {i18n.modules.card.createVariant}
-                </>
-              ),
-              action: () => {
-                dispatch(API.createCardContentVariantWithBlockDoc(card.id!)).then(payload => {
-                  if (payload.meta.requestStatus === 'fulfilled') {
-                    if (entityIs(payload.payload, 'CardContent')) {
-                      goto(card, payload.payload);
+                  ]
+                : []),
+              {
+                value: 'createVariant',
+                label: (
+                  <>
+                    <Icon icon={'library_add'} /> {i18n.modules.card.createVariant}
+                  </>
+                ),
+                action: () => {
+                  dispatch(API.createCardContentVariantWithBlockDoc(card.id!)).then(payload => {
+                    if (payload.meta.requestStatus === 'fulfilled') {
+                      if (entityIs(payload.payload, 'CardContent')) {
+                        goto(card, payload.payload);
+                      }
                     }
-                  }
-                });
+                  });
+                },
               },
-            },
-          ]}
-        />
+            ]}
+          />
+        </Flex>
       </Flex>
-    </Flex>
+      <Flex direction="column" align="stretch">
+        {readOnly ? (
+          <ProgressBar card={card} variant={cardContent} tall />
+        ) : (
+          <ProgressBarEditor card={card} variant={cardContent} />
+        )}
+      </Flex>
+    </>
   );
 }
